feat: add run-addressable create-issue eval manifests
