+++ lib/test/fixtures/create-issue-eval/manifests/runs/baseline/audit-state.json
@@ -1,0 +1,10 @@
+{
+  "rounds": [
+    {
+      "round": 1,
+      "kind": "discovery",
+      "kind_reason": "whole_draft_check",
+      "findings": []
+    }
+  ]
+}
