+++ lib/test/fixtures/create-issue-eval/manifests/runs/baseline/draft-initial.md
@@ -1,0 +1,1 @@
+# Initial baseline draft
