+++ lib/test/fixtures/create-issue-eval/manifests/runs/candidate/draft-initial.md
@@ -1,0 +1,1 @@
+# Initial candidate draft
