+++ lib/test/fixtures/create-issue-eval/manifests/runs/candidate/draft-revision-1.md
@@ -1,0 +1,1 @@
+# Candidate revision
