+++ lib/test/fixtures/create-issue-eval/manifests/runs/shared/transcript.jsonl
@@ -1,0 +1,4 @@
+{"type":"assistant","attributionSkill":"devflow:create-issue","message":{"usage":{"input_tokens":10},"content":[{"type":"tool_use","id":"dispatch-1","name":"Bash","input":{"command":"issue-audit-state.py record-dispatch --round 1"}}]}}
+{"type":"assistant","isSidechain":true,"attributionSkill":"devflow:create-issue","message":{"usage":{"input_tokens":5}}}
+{"type":"assistant","attributionSkill":"devflow:create-issue","message":{"usage":{"input_tokens":100},"content":[{"type":"tool_use","id":"dispatch-2","name":"Bash","input":{"command":"issue-audit-state.py record-dispatch --round 1"}}]}}
+{"type":"assistant","isSidechain":true,"attributionSkill":"devflow:create-issue","message":{"usage":{"input_tokens":50}}}
