+++ lib/test/fixtures/create-issue-eval/manifests/two-occurrences.json
@@ -1,0 +1,67 @@
+{
+  "schema_version": 1,
+  "root": ".",
+  "benchmark_id": "two-occurrences",
+  "runs": [
+    {
+      "run_id": "baseline-vague-1",
+      "configuration": "baseline",
+      "scenario_id": "vague-one-liner",
+      "repetition": 1,
+      "transcript": "runs/shared/transcript.jsonl",
+      "state_file": "runs/baseline/audit-state.json",
+      "occurrence": {
+        "session_id": "session-a",
+        "occurrence_id": "create-issue-1",
+        "start_event": 0,
+        "end_event": 1,
+        "boundary_confidence": "exact",
+        "duration_ms": 12000
+      },
+      "checkpoints": {
+        "initial": "runs/baseline/draft-initial.md",
+        "revisions": [],
+        "final": "runs/baseline/draft-final.md"
+      },
+      "provenance": {
+        "repo_sha": "0123456789abcdef",
+        "skill_fingerprint": "sha256:baseline",
+        "prompt_fingerprint": "sha256:vague",
+        "model": "model-a",
+        "effort": "high",
+        "output_style": "Default",
+        "provider": "local"
+      }
+    },
+    {
+      "run_id": "candidate-vague-1",
+      "configuration": "candidate",
+      "scenario_id": "vague-one-liner",
+      "repetition": 1,
+      "transcript": "runs/shared/transcript.jsonl",
+      "state_file": "runs/candidate/audit-state.json",
+      "occurrence": {
+        "session_id": "session-a",
+        "occurrence_id": "create-issue-2",
+        "start_event": 2,
+        "end_event": 3,
+        "boundary_confidence": "exact",
+        "duration_ms": 9000
+      },
+      "checkpoints": {
+        "initial": "runs/candidate/draft-initial.md",
+        "revisions": ["runs/candidate/draft-revision-1.md"],
+        "final": "runs/candidate/draft-final.md"
+      },
+      "provenance": {
+        "repo_sha": "0123456789abcdef",
+        "skill_fingerprint": "sha256:candidate",
+        "prompt_fingerprint": "sha256:vague",
+        "model": "model-a",
+        "effort": "high",
+        "output_style": "Default",
+        "provider": "local"
+      }
+    }
+  ]
+}
